+++ 1-setup/schema/university-sqlserver.sql
@@ -43,7 +43,7 @@ CREATE TABLE Student
     birth_date DATE,
     enrollment_date DATE,
     graduation_date DATE,
-    gpa FLOAT,
+    gpa FLOAT(4),
     UNIQUE (id)
 );
